// planning-idsdp-pps-frontend/src/app/main/modules/feasibility-study-management/models/feasibility-study-summary.model.ts
export class FsSummaryModel {
    title_en: string;
    title_bn: string;
    sponsoring_ministry_id: string;
    implementing_agency_id: string;
    project_objectives: string;
    estimated_proj_cost: number;
    sector_id: number;
    sub_sector_id: number;
    project_category_id: number;
    date_of_commencement: any;
    date_of_completion: any;
    projectConceptMasterId: number;
    projectConceptMasterUuid: string;
    paripatraVersionId: number;
    uuid: any;
    id: number;


}